created a new class to handle notifications.
