+++ Shared/ContentView.swift
@@ -12,7 +12,7 @@ import UserNotifications
 struct ContentView: View {
     @Environment(\.managedObjectContext) private var viewContext
     @FetchRequest(entity: Product.entity(), sortDescriptors: []) var products: FetchedResults<Product>
-   
+    var notification = CustomNotification()
     let productTypes = ["Document","Electronics","Grocery","Subscription", "Other"]
     let daysCollection = [1, 3, 7, 30]
     
@@ -85,21 +85,31 @@ struct ContentView: View {
                 .navigationBarTitle("Add New Product")
             }
             .onAppear(perform: {
-                notificationRequest()
+                notification.notificationRequest()
                 for product in products {
-                    print("name: ",product.name)
+                    print("name: ",product.getName)
                     print("expiry date:",product.ExpiryDate)
                     print("is subscribed:", product.isNotificationSet)
                     
-                    if checkExpiry(expiryDate: product.expiryDate ?? Date(), deleteDays: product.DeleteAfter, product: product) {
-                        viewContext.delete(product)
+                    let result = notification.checkExpiry(expiryDate: product.expiryDate ?? Date(), deleteAfter: product.DeleteAfter, product: product)
+                    print("result for \(product.getName) is: \(result)")
+                    switch result {
+                        case "Expired" :
+                            viewContext.delete(product)
+                        case "Near Expiry":
+                            product.isNotificationSet = true
+                        case "Undefined":
+                            break
+                        default:
+                        break
+                    }
                         do {
                             try viewContext.save()
+                            print("\(product.getName) is saved.")
                         }
                         catch {
                             fatalError(error.localizedDescription)
-                        }
-                    }
+                        }                 
                 }
                
             })
@@ -119,19 +129,6 @@ struct ContentView: View {
                 
     }
     
-    func notificationRequest() {
-            let center = UNUserNotificationCenter.current()
-            center.requestAuthorization(options: [.alert,.badge, .sound]) { success, error in
-                if success {
-                    print("all set")
-                
-                }
-                else if let error = error {
-                    print(error.localizedDescription)
-                }
-               
-            }
-        }
     
     func prepareAlertContent(title: String, message: String) {
         alertTitle = title
@@ -139,87 +136,6 @@ struct ContentView: View {
         showAlert = true
     }
    
-    func checkExpiry(expiryDate: Date, deleteDays: Int, product: Product) -> Bool {
-        let diff = Calendar.current.dateComponents([.day], from: Date(), to: expiryDate)
-            if let days = diff.day {
-                print("days:",days)
-                print(product.ExpiryDate)
-                // if today is after expiry
-                if days < 0 {
-                    if abs(days) >= deleteDays {
-                        return true
-                    }
-                }
-                // if today is before expiry
-                else {
-                    if days <= 3 {
-                       
-                       // sendNotification(product: product)
-                        if !product.isNotificationSet {
-                             sendNotification(product: product)
-                            print("calling notifcation for \(product.getName)")
-                            //localNotification.setLocalNotification (title: "\(product.getName) is expiring soon!", subtitle: "\(product.getName) expiring \(product.ExpiryDate == dateFormatter.string(from: Date()) ? "today.": "on \(product.ExpiryDate).")", body: "", when: date)
-                        }
-                       
-                        
-                    }
-                    return false
-                }
-            }
-            
-        return false
-    }
-    func sendNotification(product: Product) {
-        
-        let content = UNMutableNotificationContent()
-        content.title = "\(product.getName) is expiring soon!"
-        content.subtitle = "\(product.getName) expiring \(product.ExpiryDate == dateFormatter.string(from: Date()) ? "today.": "on \(product.ExpiryDate).")"
-        content.sound = UNNotificationSound.default
-        
-        let addRequest =  {
-            let trigger = UNCalendarNotificationTrigger(dateMatching: date, repeats: true)
-            let request = UNNotificationRequest(identifier: UUID().uuidString, content: content, trigger: trigger)
-            UNUserNotificationCenter.current().add(request) {error in
-                guard let error = error else {
-                    return
-                }
-                fatalError(error.localizedDescription)
-            }
-        }
-      
-        UNUserNotificationCenter.current().getNotificationSettings { settings in
-            if settings.authorizationStatus == .authorized {
-                print("request is added.")
-                
-                addRequest()
-                do {
-                    product.isNotificationSet = true
-                    try viewContext.save()
-                    print("product \(product.getName) saved")
-                }
-                catch {
-                    fatalError(error.localizedDescription)
-                }
-            }
-            else if settings.authorizationStatus == .notDetermined {
-                UNUserNotificationCenter.current().requestAuthorization(options: [.alert, .badge, .sound]) { success, error in
-                    print("request is not added yet.")
-                    if success {
-                        addRequest()
-                        print("request is now added.")
-                    }
-                    else {
-                        fatalError((error != nil) ? error!.localizedDescription : "Unknown Error." )
-                                                
-                    }
-                }
-            }
-            else {
-                return
-            }
-        }
-    }
-    
     func addProduct() {
         let product = Product(context: viewContext)
         product.name = productName
@@ -231,8 +147,6 @@ struct ContentView: View {
         do {
             try viewContext.save()
             print("product saved")
-           
-            
         }
         catch {
             fatalError(error.localizedDescription)
